dllテストしてみました 例外処理が弱く動作が安定しないので一度見直して整理するべき状態です。

+++ OpenAL_Engine/OpenAL/OpenAlCore/OpenALDevice.h
@@ -15,7 +15,7 @@
 namespace htAudio
 {
 	const int SAMPLE_PER_SEC = 44100;
-	const int BUFFER_SIZE = 6096;
+	const int16_t BUFFER_SIZE = 6096;
 	
 	enum AlutError
 	{
